Move ProgressBar to common
